fix(plugins): resolve dreaming sidecar exemption in shared activation state
Scope the dreaming-engine-sidecar allowlist exemption to bundled origin and
to configs where the plugin would otherwise be rejected by a restrictive
allowlist, so no other activation decision changes.

Derive the dreaming sidecar engine id centrally in resolvePluginActivationState
instead of threading it from the loader, so every registry surface that funnels
through config-state (gateway loader, CLI plugin registry, installed-plugin
index, openclaw plugins list) agrees the sidecar is enabled. Moves
resolveDreamingSidecarEngineId from loader.ts to config-state.ts.
